refactor : 플레이어 카드 리스트를 일급컬렉션으로 추상화해서 분리
